# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: GustoPOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ingredients:
    get:
      operationId: getIngredients
      tags: [ingredients]
      summary: List all ingredients/inventory
      responses:
        "200":
          description: List of ingredients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ingredient"
    post:
      operationId: createIngredient
      tags: [ingredients]
      summary: Create an ingredient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIngredientBody"
      responses:
        "201":
          description: Created ingredient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
  /ingredients/{id}:
    get:
      operationId: getIngredient
      tags: [ingredients]
      summary: Get an ingredient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ingredient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
    patch:
      operationId: updateIngredient
      tags: [ingredients]
      summary: Update an ingredient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIngredientBody"
      responses:
        "200":
          description: Updated ingredient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
    delete:
      operationId: deleteIngredient
      tags: [ingredients]
      summary: Delete an ingredient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /promo-codes/{code}:
    get:
      operationId: getPromoCodeByCode
      tags: [promo-codes]
      summary: Get promo code by code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Promo code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoCode"
  /tabs/{id}/apply-code:
    patch:
      operationId: applyPromoCodeToTab
      tags: [promo-codes]
      summary: Apply promo code to tab
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyPromoCodeBody"
      responses:
        "200":
          description: Applied promo code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplyPromoCodeResponse"
...
  /rushes:
    get:
      operationId: getRushes
      tags: [rushes]
      summary: List local events and rushes
      responses:
        "200":
          description: List of rushes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rush"
    post:
      operationId: postRushes
      tags: [rushes]
      summary: Create a new rush event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRushBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rush"
  /rushes/{id}:
    delete:
      operationId: deleteRushesId
      tags: [rushes]
      summary: Delete a rush event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Ingredient:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nameEs:
          type: ["string", "null"]
        type:
          type: string
          enum: [spirit, beer, mixer, unit]
        subtype:
          type: ["string", "null"]
        baseUnit:
          type: string
          enum: [ml, g, unit]
        baseUnitAmount:
          type: number
        servingSize:
          type: number
        sellSingleServing:
          type: boolean
        singleServingPrice:
          type: ["number", "null"]
        fullBottleWeightG:
          type: ["number", "null"]
        containerWeightG:
          type: ["number", "null"]
        density:
          type: ["number", "null"]
        currentStock:
          type: number
        currentBulk:
          type: number
        currentPartial:
          type: number
        reservedStock:
          type: number
        orderCost:
          type: number
        lowStockThreshold:
          type: number
        unitsPerCase:
          type: integer
        trackingMode:
          type: string
          enum: [auto, pool, collection]
        parentItemId:
          type: ["string", "null"]
        isOnMenu:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - baseUnit
        - baseUnitAmount
        - servingSize
        - currentStock
        - orderCost
        - lowStockThreshold
        - unitsPerCase
        - isOnMenu
        - createdAt
        - updatedAt
    CreateIngredientBody:
      type: object
      properties:
        name:
          type: string
        nameEs:
          type: ["string", "null"]
        type:
          type: string
          enum: [spirit, beer, mixer, unit]
        subtype:
          type: ["string", "null"]
        baseUnit:
          type: string
          enum: [ml, g, unit]
        baseUnitAmount:
          type: number
        servingSize:
          type: number
        sellSingleServing:
          type: boolean
        singleServingPrice:
          type: ["number", "null"]
        fullBottleWeightG:
          type: ["number", "null"]
        currentStock:
          type: number
        orderCost:
          type: number
        lowStockThreshold:
          type: number
        unitsPerCase:
          type: integer
        trackingMode:
          type: string
          enum: [auto, pool, collection]
        isOnMenu:
          type: boolean
      required:
        - name
        - baseUnit
        - baseUnitAmount
        - servingSize
        - currentStock
    UpdateIngredientBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        nameEs:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        subtype:
          type: ["string", "null"]
        baseUnit:
          type: ["string", "null"]
        baseUnitAmount:
          type: ["number", "null"]
        servingSize:
          type: ["number", "null"]
        sellSingleServing:
          type: ["boolean", "null"]
        singleServingPrice:
          type: ["number", "null"]
        fullBottleWeightG:
          type: ["number", "null"]
        currentStock:
          type: ["number", "null"]
        orderCost:
          type: ["number", "null"]
        lowStockThreshold:
          type: ["number", "null"]
        unitsPerCase:
          type: ["integer", "null"]
        trackingMode:
          type: string
          enum: [auto, pool, collection]
        isOnMenu:
          type: ["boolean", "null"]
...
    Rush:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        startTime:
          type: integer
        endTime:
          type: ["integer", "null"]
        repeatEvent:
          type: integer
          enum: [0, 1, 2, 3]
          description: "0=never, 1=weekly, 2=monthly, 3=daily"
        impact:
          type: string
          enum: [low, medium, high]
        type:
          type: string
          enum: [cruise, festival, music, other]
      required:
        - id
        - title
        - startTime
        - impact
        - type
    CreateRushBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        startTime:
          type: integer
        endTime:
          type: ["integer", "null"]
        repeatEvent:
          type: integer
          enum: [0, 1, 2, 3]
          description: "0=never, 1=weekly, 2=monthly, 3=daily"
        impact:
          type: string
          enum: [low, medium, high]
        type:
          type: string
          enum: [cruise, festival, music, other]
      required:
        - title
        - startTime
        - impact
        - type
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    PromoCode:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        description:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed_amount]
        discountValue:
          type: number
        maxUses:
          type: integer
        currentUses:
          type: integer
        expiresAt:
          type: string
          format: date-time
        isActive:
          type: boolean
      required:
        - id
        - code
        - discountType
        - discountValue
    ApplyPromoCodeBody:
      type: object
      properties:
        promoCode:
          type: string
      required:
        - promoCode
    ApplyPromoCodeResponse:
      type: object
      properties:
        id:
          type: string
        discountMxn:
          type: number
        promoCodeId:
          type: string
      required:
        - id
        - discountMxn
        - promoCodeId
...